% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_whatnot.R
\name{printnum}
\alias{printnum}
\title{popdata() function by Theo (SOC21)
This function takes input from the user on allele names and population size
It then prints out the genetic variation data}
\usage{
printnum(n)
}
\description{
popdata() function by Theo (SOC21)
This function takes input from the user on allele names and population size
It then prints out the genetic variation data
}
